requirements.txt
Django>=2.1.5
django-tastypie>=0.9.14